simplify markup
